+++ packages/yourloops/lib/team/hook.tsx
@@ -306,6 +306,15 @@ function TeamContextImpl(teamAPI: TeamAPI, directShareAPI: DirectShareAPI): Team
     return res;
   };
 
+  const editPatientRemoteMonitoring = (patient: Patient) => {
+    const user = getUser(patient.userid);
+    if (!user) {
+      throw Error("Cannot update user monitoring as user was not found");
+    }
+    user.monitoring = patient.monitoring;
+    setTeams(teams);
+  };
+
   const getMedicalMembers = (team: Team): TeamMember[] => {
     return team.members.filter((member) => member.role !== TeamMemberRole.patient);
   };
@@ -707,6 +716,7 @@ function TeamContextImpl(teamAPI: TeamAPI, directShareAPI: DirectShareAPI): Team
     markPatientMessagesAsRead,
     createTeam,
     editTeam,
+    editPatientRemoteMonitoring,
     updateTeamAlerts,
     leaveTeam,
     removeMember,